requirement.txt
Flask
mysql-connector-python
pandas
together-sdk
gunicorn
